# Performance-focused rewrite of the ConvexPolyhedron class