… .circleci/config.yml
… ChangeLog.rst
… Credits.rst
… coxeter/shapes/polyhedron.py
… doc/source/coxeter.bib
… tests/conftest.py
… tests/test_polyhedron.py