Prior to this PR, the ConvexPolyhedron class used Scipy's (Qhull) convex hull to compute faces of the solid. The shape was then initialized in the same manner as a standard Polyhedron. This rewrite stores simplices and other data from the convex hull, which can be used to accelerate many of the classes methods. Slow and commonly run functions have been rewritten for performance, and new and/or improved pytests have been written.

## Description
The new class builds a convex hull, then stores most of the data from that object. This data, especially the list of simplices, allow for numpy vectorization of many functions that required a for loop in the previous code. These simplices can be sorted to ensure math involving the curl/divergence theorems works properly, and can be merged based on common normals to build faces similar to the current merge_faces function. 


**The following methods have been reimplemented, and a brief description is given:**

* `find_equations`: vectorized using numpy
* `_surface_triangulation`: updated to use precalculated simplices, so polytri is not required
* `volume`, `surface_area`: reimplemented to use stored volume data for the calculations. 
* `centroid`: rewritten to calculate from a list of simplices. Vectorized using numpy, and computes solutions using (`_centroid_from_triangulated_surface`)
* `_compute_inertia_tensor`: fully rewritten for performance, using an [algorithm](https://dl.acm.org/doi/10.1145/355873.355885) that works with tetrahedral decompositions that include negative-volume tetrahedra. It is also significantly faster.
* `_rescale`: updated to properly scale all stored properties. Also recalculates the centroid to ensure proper behavior.
* `get_face_area`: vectorized with numpy

**The following new methods have been added:**

* `_combine simplices`: faster alternative to merge faces, but only works on simplices
* `_find_simplex_equations`: similar to _find_equations, but works on every simplex rather than every face
* `_find_coplanar_simplices`: helper function that finds simplices with the same plane equation (within a tolerance). 
* `_calculate_signed_volume`: faster method of computing volume via a tetrahedral decomposition. Used as an internal check that volume is stable when modifying the shape
* `_find_face_centroids`: used with the new face_centroids property. Quickly finds the center of mass for each of the polygonal faces

**The following properties have been added:**
* `simplices`
* `equations`
* `face_centroids`

## Motivation and Context
Creation of a ConvexPolyhedron object takes on the order of 1e-2s - this is fine when generating small numbers of shapes, but can become a significant hindrance when generating tens or hundreds of thousands of shapes. In addition, many commonly used methods (especially volume, centroid, and inertia tensor) took on the order of 1e-3s and did not cache their data. These changes address these performance issues, and add a wide range of private variables which can be used to more quickly calculate geometric properties that have not yet been implemented. Benchmarks are included below, but in general this rewrite provides a 10x-100x performance improvement on improved methods. For properties with caching, values are only recalculated when the shape is modified and access the data takes on the order of 1e-8s

## Benchmarks
_Note that tests were performed in much more depth than shown below - these are simply a subset of the total_

### With optimization

#### Cube (V=8, F=6, E=12)
```python
cube = PlatonicFamily.get_shape("Cube")

%timeit -r 10 -n 10_000 ConvexPolyhedron(cube.vertices)
> 301 us ± 34.2 us per shape

%timeit -r 10 -n 100_000 cube._calculate_signed_volume()
> 7.69 us ± 239 ns per shape

%timeit -r 10 -n 100_000 cube._centroid_from_triangulated_surface()
> 25.7 us ± 676 ns per shape

%timeit -r 10 -n 100_000 cube.inertia_tensor
> 111 us ± 2.74 us per shape
```

#### Metabidiminished rhombicosidodecahedron (V=50, F=42, E=90)
```python
mbri = JohnsonFamily.get_shape("Metabidiminished Rhombicosidodecahedron")

%timeit -r 10 -n 10_000 ConvexPolyhedron(mbri.vertices)
> 1.81 ms ± 315 us per shape

%timeit -r 10 -n 100_000 mbri._calculate_signed_volume()
> 20.6 us ± 2.38 us per shape

%timeit -r 10 -n 100_000 mbri._centroid_from_triangulated_surface()
> 32.2 us ± 1.96 us per shape

%timeit -r 10 -n 100_000 mbri.inertia_tensor
> 194 us ± 9.65 us per shape
```

### Without optimization (previous versions of coxeter)

#### Cube (V=8, F=6, E=12)
```python
cube = PlatonicFamily.get_shape("Cube")

%timeit -r 10 -n 10_000 ConvexPolyhedron(cube.vertices)
> 5.22 ms ± 169 us per shape

%timeit -r 10 -n 100_000 cube.volume
> 1.38 ms ± 22.8 us per shape

%timeit -r 10 -n 100_000 cube.centroid
> 946 us ± 37.4 us per shape

%timeit -r 10 -n 100_000 cube.inertia_tensor
> 5.32 ms ± 478 us per shape
```

#### Metabidiminished rhombicosidodecahedron (V=50, F=42, E=90)
```python
mbri = JohnsonFamily.get_shape("Metabidiminished Rhombicosidodecahedron")

%timeit -r 10 -n 1_0000 ConvexPolyhedron(mbri.vertices)
> 38.1 ms ± 1.05 ms per shape

%timeit -r 10 -n 10_000 mbri.volume
> 10.2 ms ± 328 us per shape

%timeit -r 10 -n 10_000 mbri.centroid
> 8.34 ms ± 829 us per shape

%timeit -r 10 -n 10_000 mbri.inertia_tensor
> 34.5 ms ± 1 ms per shape
```

## Types of Changes
- [ ] Documentation update
- [ ] Bug fix
- [x] New feature
- [x] Breaking change<sup>1</sup>

<sup>1</sup>The change breaks (or has the potential to break) existing functionality.

## Checklist:
- [x] I am familiar with the [**Contributing Guidelines**](https://github.com/glotzerlab/coxeter/blob/master/CONTRIBUTING.md).
- [x] I agree with the terms of the [**Contributor Agreement**](https://github.com/glotzerlab/coxeter/blob/master/ContributorAgreement.md).
- [x] The changes introduced by this pull request are covered by existing or newly introduced tests.
- [x] I have updated the [changelog](https://github.com/glotzerlab/coxeter/blob/master/ChangeLog.txt) and the [credits](https://github.com/glotzerlab/coxeter/blob/master/doc/source/credits.rst).