@janbridley 2023-08-09
[0cf79cb](https://github.com/glotzerlab/coxeter/pull/188/commits/0cf79cb9b78ad40b6eaa4fe2cb02183fff490c40) is failing because it is dependent on #171 . Once that PR is merged, tests will pass