@janbridley 2023-08-29
> there are a lot of design questions here (such as whether it is a good idea to add a flag like fast=True to the constructor, which makes this class behave quite differently than  the rest of coxeter). Generally, coxeter was designed to provide a common framework for  representing geometry classes (which we previously lacked in the Glotzer group), and speed was not the central goal. This design choice is reflected in its implementation (aiming for readability and simplicity of implementation over speed in most cases).


Totally valid to not have the bandwidth this week - I looped in @DomFijan for this. 

In terms of the scope of coxeter and the differences in mindset between this code (and a few other recent PRs) as compared to the state of the package a year ago: coxeter was created as a common geometry framework, but it was not initially intended to be used beyond creating a few dozens of shapes at a time. However, the tools here are widely applicable and many glotzergroup members (and presumably a few other researchers) are using the package as a comprehensive computational geometry package. This includes use creating hundreds or thousands of shapes at a time, as well as using coxeter inside molecular dynamics scripts to easily compute shape properties. There is definitely a use case for faster algorithms, and I would argue that although simple iteration based code may be easier to read, the barrier to understanding reasonably optimized numpy code is not significantly higher than the alternative. Speed is not the primary purpose of coxeter, and I do not think it should be - but orders-of-magnitude performance improvements with no loss in accuracy are well worth the slight increase in complexity.


In terms of the `fast` flag in particular: although its effectiveness is somewhat limited in the current implementation, there are plans to support convex decompositions in the future which would benefit from being able to use some or all of the ConvexPolyhedron code without a large number of extra calculations being performed on each segment. That being said, I am fine removing the feature in this PR, as it is not particularly difficult to implement.