@janbridley 2023-09-27
I have marked this PR as a draft temporarily, as there is a significant bug when creating a ConvexPolyhedron with a non-convex input set. A failure in this case is fairly reasonable, but the error should be (a) more predictable and (b) well-documented.

### Current behavior:
```python
>>> verts
array([[-0.5,  0.5, -0.5],
       [-0.5,  0.5,  0.5],
       [ 0.5, -0.5, -0.5],
       [-0.5, -0.5, -0.5],
       [ 0.5,  0.5,  0.5],
       [-0.5, -0.5,  0.5],
       [ 0.5,  0.5, -0.5],
       [ 0. ,  0. ,  0. ],
       [ 0.5, -0.5,  0.5]])
>>> coxeter.shapes.ConvexPolyhedron(verts)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/$USER/coxeter/coxeter/shapes/convex_polyhedron.py", line 114, in __init__
    self._sort_simplices()
  File "/Users/$USER/coxeter/coxeter/shapes/convex_polyhedron.py", line 205, in _sort_simplices
    vertices = self._vertices[self._simplices]
               ~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^
IndexError: index 8 is out of bounds for axis 0 with size 8
```

When a non convex set is used as an input, vertices that don't contribute to the hull are removed from self._vertices. However - the _indices_ to those vertices stored in self._simplices are not updated, raising this error.