@janbridley 2023-10-23
The class has been updated to raise a ValueError when a nonconvex set is provided. This error will also return a list of internal (or coplanar, degenerate, etc) points for easier debugging.